requirements_test.txt
antlr4-python3-runtime>=4.11
datrie>=0.8.2
docutils>=0.18.1
grammar_graph>=0.1.14
grammar_to_regex>=0.0.4
ijson>=3.1.4
matplotlib>=3.5.3
packaging>=21.3
pathos>=0.2.9
proxyorderedset>=0.3.0
pytest-cov>=3.0.0
pytest-html>=3.1.1
pytest-profiling>=1.7.0
pytest-pycharm>=0.7.0
pytest-randomly>=3.12.0
pytest-rerunfailures>=10.2
pytest-xdist>=2.4.0
pytest>=7.1.2
setuptools-antlr>=0.4.0
tox>=3.25.0
z3-solver>=4.8.17.0
